Rename paths -> runtime_environment
